fix(ux): consolidate Convert into a single coral chip
The Toolbar carried its own Convert/Stop button (and the ⌘/Ctrl+Enter
shortcut) while a floating ConvertChip rendered the same action over the
editor / preview divider. Two buttons, identical action — confusing.

Drop the toolbar pair, leave the chip as the single source of Convert
truth, and migrate the keyboard shortcut onto it. Recolor the chip from
the previous near-black `var(--ink)` to `var(--coral)` (idle) /
`var(--coral-hover)` (running) so the primary CTA is visually distinct
from the muted toolbar surface and from text content. Disabled state
still renders coral at 0.4 opacity.
